hotfix: officialOnly parsing bug with sets
