+++ .github/workflows/starrailoptimizer.yml
@@ -8,8 +8,8 @@ jobs:
   build:
     runs-on: ubuntu-latest
     container: pandoc/latex    # "ubuntu" is a more generic container
-                               # using "pandoc/latex" because of dependencies
-                               # used in the specific "build.sh"
+    # using "pandoc/latex" because of dependencies
+    # used in the specific "build.sh"
     steps:
       - uses: actions/checkout@v4
       - name: Find and Replace homepage
@@ -39,9 +39,9 @@ jobs:
       - name: Find and Replace officialOnly
         uses: jacobtomlinson/gha-find-replace@v3
         with:
-          find: "window.officialOnly = false"
-          replace: "window.officialOnly = true"
-          include: "src/index.jsx"
+          find: "export const officialOnly = false"
+          replace: "export const officialOnly = true"
+          include: "src/lib/constants.ts"
       - run: rm -rf .github/workflows
       - run: mv .github/starrailoptimizer .github/workflows
       - name: Pushes to another repository
