+++ src/components/ChangelogTab.tsx
@@ -3,6 +3,7 @@ import { Flex, List, theme, Typography } from 'antd'
 import { AppPages } from 'lib/db.js'
 import { Assets } from 'lib/assets'
 import { ColorizedLink } from './common/ColorizedLink'
+import { officialOnly } from 'lib/constants'
 
 const { useToken } = theme
 const { Text } = Typography
@@ -95,7 +96,7 @@ export default function ChangelogTab(): React.JSX.Element {
 }
 
 function leaks(str: string) {
-  return window.officialOnly ? '' : str
+  return officialOnly ? '' : str
 }
 
 /*
