+++ src/components/MenuDrawer.jsx
@@ -5,6 +5,7 @@ import { DiscordIcon } from 'icons/DiscordIcon'
 import { GithubIcon } from 'icons/GithubIcon'
 import { CoffeeIcon } from 'icons/CoffeeIcon'
 import { AppPages } from 'lib/db'
+import { officialOnly } from 'lib/constants'
 
 function getItem(label, key, icon, children, type) {
   return {
@@ -123,7 +124,7 @@ const MenuDrawer = () => {
         </Typography.Link>,
         'link github',
       ),
-      !window.officialOnly && getItem(
+      !officialOnly && getItem(
         <Typography.Link href='https://starrailoptimizer.github.io/' target='_blank' rel='noopener noreferrer'>
           <LinkOutlined style={{ marginRight: 5 }}/>
           {' '}
