+++ src/components/RelicFilterBar.jsx
@@ -7,14 +7,13 @@ import { TooltipImage } from './TooltipImage'
 import DB from '../lib/db'
 import { Hint } from 'lib/hint'
 import { Utils } from 'lib/utils'
-import { Constants, SetsRelics, Stats } from 'lib/constants'
+import { Constants, SetsRelics, Stats, UnreleasedSets } from 'lib/constants'
 import { Assets } from 'lib/assets'
 import PropTypes from 'prop-types'
 import { useSubscribe } from 'hooks/useSubscribe'
 import { Renderer } from 'lib/renderer'
 import CharacterSelect from 'components/optimizerTab/optimizerForm/CharacterSelect'
 import { ClearOutlined } from '@ant-design/icons'
-import { UnreleasedSets } from 'lib/dataParser'
 import { SaveState } from 'lib/saveState'
 import { SettingOptions } from 'components/SettingsDrawer'
